Testing SQLite and Zoom
Tested out SQLite4Unity3D for “chaining” shots (spawning an object
after taking a shot of another, allowing for a sequence of shots). Also
tested out spawning objects for a limited time when you gaze at
another, for a “whackamole”-like game mechanic, and the use of zoom
when the player holds down the trigger.