Added win and lose panels, removed bonus roll feature
Added panels which display for each of the different win and lose game
conditions that allow the player to restart the game. Added a pause
panel that allows the player to pause the game on tilt and that pauses
the game on soft close. Removed bonus roll feature b/c it’s redundant
with a points system.